mb/supermicro/x11-lga1151v2-series: Add support for X11SCH-F
This is a µATX UP server mainboard with a LGA1151 socket and four DDR4
DIMM slots. Unlike existing boards, this one supports Coffee Lake CPUs
instead of Skylake and Kaby Lake, thus it is not added as a variant.

Tested with an Intel Xeon E-2276G CPU and 4x 16 GiB ECC DDR4 UDIMMs.

Working:
- All four DIMM slots
- Aspeed AST2500 graphics output
- Serial console
- Super I/O devicetree config
- SATA ports
- USB ports
- IPMI KCS
- BMC iKVM (keyboard/mouse input not available before OS loads)
- TianoCore from 9elements/edk2
- LinuxBoot
- Booting Linux 5.5 and Windows 10
- No ACPI errors in Windows/Linux
- TPM support (w/ patched IFD) - Patch IFD at 0x235=0x03 and 0x13E=0x84

Untested:
- PCIe/M.2 ports and corresponding SMBIOS information

Change-Id: I0ab1cb9462607b9af068bc2374508d99c60d0a30
Signed-off-by: Christian Walter <christian.walter@9elements.com>